requirements.txt
black == 23.12.1
jsonschema == 4.20.0
pre-commit == 3.6.0
pyyaml == 6.0.1